update to float64

+++ ai_agents/v2/gym/full_information_protagonist_antagonist_gym.py
@@ -14,8 +14,8 @@ TABLE_MAX_Y_DIM = 65
 BALL_STOPPED_COUNT_THRESHOLD = 10
 MAX_STEPS = 40
 SIM_PATH = os.environ.get('SIM_PATH', '/home/ppy2104/foosballpart2/foosball_sim/v2/foosball_sim.xml')
-F32   = np.float32
-INF32 = np.finfo(np.float32).max
+F32   = np.float64
+INF32 = np.finfo(np.float64).max
 
 RODS = ["_goal_", "_def_", "_mid_", "_attack_"]
